addNodeAction check if action already exists
In order to avoid duplicate Node mappings (since matchEdge in
Match.hs automatically adds the source/target Nodes to the
Morphism), addNodeAction checks if this nodes are already
contained in the given morphism.

I wanted to just compare the Nodes, since they derive the Eq type
class, but ghc complained that a/b also should be an instance of Eq.
If I understand it right, this would require many changes in other
functions from Match.hs. So I did this hack to compare the ID's.
Just in case, the old code is commented out.